Update from nopt 4.x to nopt 5.0.0 (#156)
installation of latest grunt-cli@1.4.2 produces a warning:

```
npm WARN deprecated osenv@0.1.5: This package is no longer supported.
```

This package is only used by nopt to call `os.homedir()` which has
been stable since Node.js 2. It was removed [1] in nopt 5.0.0.

There are newer versions of nopt available, but those raise the
required Node.js engine level. In order to make this easy to release
and to resolve the warning first, let's first move to nopt 5.0.0.

[1]: https://github.com/npm/nopt/commit/5c0e45b2f94414c074270b133e90c77ec35f5a4b